Add rubocop test case with un-corrected offense
If an offense is found and not corrected, it will cause a non-zero exit.
This is undesired. Which offenses lead to this behavior can be
controlled with --fail-level, but it's unclear if there is a level we
can provide that means "never exit non-zero". I'm trying "fatal" for
now, which works for the known example at least.