Tweak build dependencies
- Need only pushed for manifest
- Need only tested for pushing

This means that if I edit a test, the push will re-fire, but it'll just
push an existing image as a new tag and rewrite the manifest. This may
be undesired eventually, but for now it's kind of a nice way to update
the manifest naturally since we've moved to explicit tags.