Don't consider the whole tracked directory
Because this called getDirectoryFiles, that meant that untracked files
appearing would cause a rebuild. And Dockerfile.* are new files the
first time we introduce a restyler, which meant we'd see an extra
rebuild and dirtied manifest.

At present, we have no files other than the Dockerfile for each
restyler, and it'd be nice to keep it that simple for many reasons. If
we can do that, we only have to "need" the Dockerfile itself. Avoiding
git-status in the build dependencies is easier to understand in a lot of
ways, including this one.