Update manifest for unnecessary image churn
Something got out of whack with the shake DB and lots of images rebuilt.

Hopefully moving to a digest-based Change option will prevent this in
the future, but we still need to commit the shas produced by that
previous churn. We will not update Restyler for these since we expect no
actual changes to have occurred.