Supply saner options to shfmt
We try to avoid doing this as much as possible, but the defaults here
are just too surprising to shell programmers. The author of shfmt is a
go developer and chose to set defaults to make the shell be like go. We
think it makes more sense for the shell to be like shell.