Add support for override Restylers
An "override" Restyler is one that uses the image (and any other
attributes) and an existing Restyler, but has its own name and possibly
other options (such as arguments). This is useful for something like
prettier where the single image can handle a multitude of languages, but
we'd like to treat it as distinct Restylers with specific includes or
behavior.

Info is first pulled from the "parent" named as an override, then any
keys there will be merged over that. It's expected that this will
specify a pre-existing image, so there is no machinery for building or
pushing anything. The Makefile handles testing only and a test/$name.t
file is expected.