Remove diff noise from test suite
This diff noise doesn't matter to the Restyler and having to add it is
annoying when implementing new Restylers. Now, all that should appear in
the expectation are the changed lines.

In a multi-file/multi-hunk case, it may be a bit surprising to see the
hunks smashed together, we'll see, but for the simple, single-hunk,
"this Restyler doesn't crash" examples, it's very nice.