Fix rustfmt when rust-toolchain file is present
The presence of a ./rust-toolchain file with a different toolchain than
what we've used here will cause rustfmt to attempt to sync and reinstall
itself from that toolchain with rustup. This of course fails because of
no networking.

This commit reproduces and then resolves that error. And updates to
rustfmt-1.4.14 along the way.

Closes #196.