Smooth new releases
We want to ensure things can be pulled before release. When there's a
new version we need to build/push it. It's annoying to remember to do
that separately, so now release will try it on any exceptions pulling.

It's annoying to pre-check for existence, so we'll just do this on any
error and then try to re-pull again after to make sure it was
appropriate after the fact.