Declare google-java-format version in a build.gradle
So renovate-bot will manage it for us.

Attempting to actually install an executable from this build file was a
non-starter. Perhaps if I knew Java I could make it work, but-- even
after figuring out how to get gradle working at all --it just wanted to
get things on a classpath for a "project", and not install an executable
dependency for me.

In the end, we just sed-awk-parse out the version from the gradle file
and continue the same executable download as we were -- only now,
changing that build.gradle file (which renovate-bot will do) will cause
us to download a newer executable. Huzzah.