Update Fourmolu, use pre-built binaries
The project has started hosting binaries in its releases. This means we
don't have to build it, which is both simpler and means we don't can use
the latest version without figuring out a ghc-9.6 build to produce it.

I'm assuming that a Fourmolu built with newer GHC can still reformat
code targeting an older GHC, but not necessarily the reverse. This
should make it safe (and best) to use the latest version we can to cover
the widest range of projects we can.