Add clazy restyler
I'm not 100% this is working since I can't get it to do anything without
using `-checks=level2`. The behavior doesn't seem to match the
documentation there, which could be a me problem or not.

In any event, this is not harmful and so is good to ship, now that I was
able to add that `-checks=level2` in only the tests. Maybe an
experienced user can figure this out.