Ensure clang-format is tagged with v-prefix
We never had a reason to care for including the `v` prefix or not, so
we've been inconsistent so far. However, when implementing the "series"
image feature, we are going to get more strict and only generate series
images (`vX`, `vX.Y`) if the tag strictly matches a `vX.Y.Z` semantic
version.

As a `clang-format` user is requesting that feature, and it just happens
to have been one of the non-`v` cases, we have to change it.

This could cause things to sort oddly when compared, `v16.0.1` vs
`16.0.1`, but I don't foresee any risk to actual functionality, or know
of any place we do such comparisons.