Initial pre-commit restyler
`pre-commit` refuses to run unless in a git repository, which seems like
something that could be avoided for commands that don't interact with
git, like `install-hooks` and `run`. I could find no way around this,
and it's usually the case under normal operation that we'll be in a git
repository anyway, so I didn't want to do too much to work around it
(e.g. forking). Instead, I added a custom entrypoint that initializes
git if necessary.

We pre-install a very limited set of hooks right now. We cannot rely on
the behavior of installing hooks lazily because restylers don't have
network. But we can pre-install whatever we want during the docker
build. As of this commit, we only pre-install two versions of black.
This is just enough to prove things function, and that multiple versions
of the same hook can co-exist and be used by users, which I was worried
about.

Obviously, this is not useful yet compared to just using our black
restyler, but hopefully it's now obvious and trivial to extend it such
that it is. I'm hoping a pre-commit user could take it from here to
decide on a set of hooks and revisions that we should launch with.

NOTE: since this is such a picky auto-formatter, and it operates on
files of any extension, I think it will always be an `enabled:false`,
opt-in restyler.