cmake-format: include optional PyYAML dependency
PyYAML is an optional dependency of cmake-format, which means that it will not
automatically be installed but triggers import errors if a `cmake-format.yaml`
file is present in the project. It seems that PyYAML seems to have been included
in the base image before so that this did not cause problems.