Stop running cabal-fmt automatically
This is a bit selfish, but all of my Haskell projects use hpack and so
the cabal file is auto-generated and should not be touched. Having
Restyled fail on every push because it overwrites hpack's output is too
disruptive.

I'll consider ways to bring this back as enabled by default but also
aware enough of hpack to not run when it's in use.