Correct multiple-path handling in pyment
It looks like `pyment` usage is one path at a time, so when multiple
files are restyled at once, it fails. We can address this by setting
`supports_multiple_paths` to `false`, causing Restyler to run it once
per path. And we duplicate the existing test case to trigger the bug
before fixing it. We should probably ensure all restylers have more than
1 test case, but I'm deferring that for now.