Rewrite CI approach
Rather than test everything all the time, we now only test changed
restylers. The problem has always been producing a complete manifest out
of this process, which we solve by downloading and merging into `dev`,
only updating the changed (and tested) Restylers, but including all the
rest as-is.

Here's hoping it works on `main`!