Ensure restylers are tested on only their test files
Testing multiple restylers at once, that operate on the same files, were
operating on each others test files and causing failures.

To address this, we take the files created for each restyler and then
put only them as the configured `include` for that restyler when the
tests are run. This ensures isolation.

It's possible we still have isolation bugs with shebang-based restyling,
but we'll cross that bridge when it causes a failure.