feat(clazy): install from git, update to 1.15
The deb datasource doesn't work because it offers versions not actually
available in the Ubuntu distribution we're using. This is despite the
renovate docs indicating the `deb` datasource should be used like this
for Debian _based_ distributions, like Ubuntu.

I can't figure out how that is actually meant to work, so I'm just
working around it by building from source. This required a few other
changes:

1. Update to 1.15, because 1.11 won't build
2. Bump ubuntu itself to 25.04, because 1.15 needs newer LLVM
3. Change the version command, now that it's not `apt`-installed