fix(prettier): change directory instead of yarn-link
Our current approach to prettier plugins was to `yarn link` them into
the current directory (`/code/node_modules`).

In the case where a user didn't commit their `node_modules` directory,
this made those plugins available for use when they otherwise would not
be. In the case where a user _did_ commit their `node_modules`, this was
probably unnecessary, since those plugins would already be present
there, but it was no harmful.

This approach was brittle and depended on "old yarn" being used by the
project. Mixing our installation with the user's package metadata (and
possible existing installation) was a recipe for problems. But it did
mean that for users with other plugins in a committed `node_modules`, we
would see/use them, as would be expected.

A simpler and more isolated approach is to run prettier from within our
installation (`/app/node_modules`) and modifying paths (presumably
relative to `/code`) to be absolute. Luckily, prettier handles executing
on files outside of the current directory correctly, including walking
up to any `.prettierrc` files, it seems.

Unfortunately, this means any committed `node_modules` will now be
ignored and so it's possible we will break any users who are relying on
that to provide plugins from outside of our blessed list.

When we find this, we could simply add the plugin to our list, but that
can only go so far. If we get requests for conflicting versions, or if
the plugin is private, respecting the committed `node_modules` may be
the only way to make that use-case work reliably.

At such time, we can consider bringing that feature back under this new
paradigm, most likely by checking for that case and doing nothing, since
executing `prettier` as-is in `/code` should _Just Work_.